+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+### Added
+- Uniform `read` and `write` methods
+
 ## [4.1.6] - 2017-06-12
 ### Fixed
 - Broken uniform setter in 4.1.5